chain_manager

+++ chain/nuls-chain/src/main/java/io/nuls/chain/service/impl/AssetServiceImpl.java
@@ -1,14 +1,15 @@
 package io.nuls.chain.service.impl;
 
 import io.nuls.base.data.chain.Asset;
+import io.nuls.base.data.chain.Chain;
 import io.nuls.base.data.chain.ChainAsset;
 import io.nuls.chain.info.CmRuntimeInfo;
 import io.nuls.chain.service.AssetService;
 import io.nuls.chain.storage.AssetStorage;
 import io.nuls.chain.storage.ChainAssetStorage;
+import io.nuls.chain.storage.ChainStorage;
 import io.nuls.tools.core.annotation.Autowired;
 import io.nuls.tools.core.annotation.Component;
-import io.nuls.tools.log.Log;
 import io.nuls.tools.thread.TimeService;
 
 import java.util.List;
@@ -27,6 +28,9 @@ public class AssetServiceImpl implements AssetService {
     @Autowired
     private ChainAssetStorage chainAssetStorage;
 
+    @Autowired
+    private ChainStorage chainStorage;
+
     /**
      * Save asset
      *
@@ -34,21 +38,26 @@ public class AssetServiceImpl implements AssetService {
      * @return true/false
      */
     @Override
-    public boolean saveAsset(Asset asset) {
-        try {
-            assetStorage.save(asset.getAssetId(), asset);
+    public boolean newAsset(Asset asset) {
+        String key = CmRuntimeInfo.getAssetKey(asset.getChainId(), asset.getAssetId());
 
-            ChainAsset chainAsset = new ChainAsset();
-            chainAsset.setChainId(asset.getChainId());
-            chainAsset.setAssetId(asset.getAssetId());
-            chainAsset.setCurrentNumber(asset.getInitNumber());
-            String key = CmRuntimeInfo.getAssetKey(chainAsset.getChainId(), chainAsset.getAssetId());
-            chainAssetStorage.save(key, chainAsset);
+        boolean s1 = assetStorage.save(asset.getAssetId(), asset);
 
+        Chain chain = chainStorage.load(asset.getChainId());
+        ChainAsset chainAsset = new ChainAsset();
+        chainAsset.setChainId(asset.getChainId());
+        chainAsset.setAssetId(asset.getAssetId());
+        chainAsset.setCurrentNumber(asset.getInitNumber());
+        chainAsset.setChainName(chain.getName());
+        chainAsset.setAssetSymbol(asset.getSymbol());
+        chainAsset.setAssetName(asset.getName());
+        boolean s2 = chainAssetStorage.save(key, chainAsset);
+
+        if (s1 && s2) {
             return true;
-        } catch (Exception e) {
-            Log.error(e);
+        } else {
             assetStorage.delete(asset.getAssetId());
+            chainAssetStorage.delete(key);
             return false;
         }
     }
